First attempt to refactor scheduler
