+++ Source/EasyNetQ/Preconditions.cs
@@ -231,5 +231,12 @@ namespace EasyNetQ
                 return;
             throw new ArgumentOutOfRangeException(name, string.Format("Arguments {0} must be less than maxValue", name));
         }
+
+        public static void CheckNull<T>(T value, string name)
+        {
+            if (value == null)
+                return;
+            throw new ArgumentException(string.Format("{0} must not be null", name), name);
+        }
     }
 }
